requirements.txt
certifi
chardet
click
Flask
idna
itsdangerous
Jinja2
joblib
jsonify
MarkupSafe
numpy
pandas
os
requests
scikit-learn
scipy
sklearn
threadpoolctl
tensorflow
keras
imutils
Pillow
matplotlib
Keras-Applications
Keras-Preprocessing
requests
opencv-python
urllib3
Werkzeug
wincertstore
gunicorn
